chore: move handleMatchmakingThreadMessage to matchmaking-threads.ts
